+++ src/check/Applications.ps1
@@ -109,22 +109,28 @@ function Invoke-InstalledApplicationPermissionCheck {
         [UInt32] $BaseSeverity
     )
 
-    process {
-        $AllResults = @()
+    begin {
+        $FsRedirectionValue = Disable-Wow64FileSystemRedirection
+    }
 
-        Get-InstalledApplication |
-        Select-Object -ExpandProperty Location |
-        Where-Object { -not [String]::IsNullOrEmpty($_) } |
-        Invoke-CommandMultithread -InitialSessionState $(Get-InitialSessionState) -Command "Get-ModifiableApplicationFile" -InputParameter "Path" |
-        ForEach-Object {
-            $_.Permissions = $_.Permissions -join ', '
-            $AllResults += $_
+    process {
+        $Results = @()
+        $Candidates = Get-InstalledApplication | Select-Object -ExpandProperty Location | Where-Object { -not [String]::IsNullOrEmpty($_) }
+        foreach ($Candidate in $Candidates) {
+            Get-ModifiableApplicationFile -Path $Candidate | ForEach-Object {
+                $_.Permissions = $_.Permissions -join ', '
+                $Results += $_
+            }
         }
 
         $CheckResult = New-Object -TypeName PSObject
-        $CheckResult | Add-Member -MemberType "NoteProperty" -Name "Result" -Value $AllResults
-        $CheckResult | Add-Member -MemberType "NoteProperty" -Name "Severity" -Value $(if ($AllResults) { $BaseSeverity } else { $script:SeverityLevel::None })
+        $CheckResult | Add-Member -MemberType "NoteProperty" -Name "Result" -Value $Results
+        $CheckResult | Add-Member -MemberType "NoteProperty" -Name "Severity" -Value $(if ($Results) { $BaseSeverity } else { $script:SeverityLevel::None })
         $CheckResult
+    }
+
+    end {
+        Restore-Wow64FileSystemRedirection -OldValue $FsRedirectionValue
     }
 }
 
@@ -331,15 +337,9 @@ function Invoke-RunningProcessCheck {
 
                 $Result = $Process | Select-Object -Property Id, SessionId, Name, Path | Add-Member -MemberType "NoteProperty" -Name "User" -Value $ProcessUser -PassThru
 
-                $Resolved = Resolve-ModulePath -Name $Process.Name
+                $Resolved = Resolve-ModuleSearchPath -Name $Process.Name -Extension ".exe"
                 if ($Resolved) {
                     $Result.Path = $Resolved
-                }
-                else {
-                    $Resolved = Resolve-ModulePath -Name "$($Process.Name).exe"
-                    if ($Resolved) {
-                        $Result.Path = $Resolved
-                    }
                 }
 
                 $Results += $Result
@@ -364,10 +364,12 @@ function Invoke-RootFolderPermissionCheck {
         # $IgnoredRootFolders = @( "Windows", "Users", "Program Files", "Program Files (x86)", "PerfLogs")
         $IgnoredRootFolders = @( "`$Recycle.Bin", "`$WinREAgent", "Documents and Settings", "PerfLogs", "Program Files", "Program Files (x86)", "ProgramData", "Recovery", "System Volume Information", "Users", "Windows", "Windows.old" )
         # $MaxFileCount = 8
-        $AllResults = @()
+        # $AllResults = @()
     }
 
     process {
+        $Results = @()
+
         # List all "fixed" drives.
         # From: https://superuser.com/a/787643
         $FixedDrives = [System.IO.DriveInfo]::GetDrives() | Where-Object { $_.DriveType -eq 'Fixed' }
@@ -378,72 +380,19 @@ function Invoke-RootFolderPermissionCheck {
             # is then filtered to exclude known folders such as "C:\Windows".
             $RootFolders = Get-ChildItem -Path $FixedDrive -Force -ErrorAction SilentlyContinue | Where-Object { ($_ -is [System.IO.DirectoryInfo]) -and ($IgnoredRootFolders -notcontains $_.Name) }
             if ($null -eq $RootFolders) { continue }
-            # foreach ($RootFolder in $RootFolders) {
 
-            #     $Vulnerable = $false
-
-            #     # Check whether the current user has any modification right on the root folder.
-            #     $RootFolderModifiablePaths = Get-ModifiablePath -Path $RootFolder.FullName | Where-Object { $_ -and (-not [String]::IsNullOrEmpty($_.ModifiablePath)) }
-            #     if ($RootFolderModifiablePaths) {
-            #         $Description = "The current user has modification rights on this root folder."
-            #     }
-            #     else {
-            #         $Description = "The current user does not have modification rights on this root folder."
-            #     }
-
-            #     # Check whether the current user has any modification right on a common app
-            #     # file within this root folder.
-            #     $ApplicationFileModifiablePaths = @()
-            #     $ApplicationFiles = Get-ChildItem -Path $RootFolder.FullName -Force -Recurse -ErrorAction SilentlyContinue | Where-Object { ($_ -is [System.IO.FileInfo]) -and (Test-IsCommonApplicationFile -Path $_.FullName) }
-            #     foreach ($ApplicationFile in $ApplicationFiles) {
-            #         if ($ApplicationFileModifiablePaths.Count -gt $MaxFileCount) { break }
-            #         if ([String]::IsNullOrEmpty($ApplicationFile.FullName)) { continue }
-            #         $ModifiablePaths = Get-ModifiablePath -Path $ApplicationFile.FullName | Where-Object { $_ -and (-not [String]::IsNullOrEmpty($_.ModifiablePath)) }
-            #         if ($ModifiablePaths) { $ApplicationFileModifiablePaths += $ApplicationFile.FullName }
-            #     }
-
-            #     # If at least one modifiable application file is found, consider the folder as
-            #     # 'vulnerable'. Even if application files are not modifiable, consider the folder
-            #     # as 'vulnerable' if the current user has any modification right on it.
-            #     if ($ApplicationFileModifiablePaths) { $Vulnerable = $true }
-            #     if ($ApplicationFiles.Count -gt 0 -and $RootFolderModifiablePaths) { $Vulnerable = $true }
-
-            #     if ($ApplicationFiles.Count -gt 0) {
-            #         if ($ApplicationFileModifiablePaths) {
-            #             $Description = "$($Description) A total of $($ApplicationFiles.Count) common application files were found. The current user has modification rights on some, or all of them."
-            #         }
-            #         else {
-            #             $Description = "$($Description) A total of $($ApplicationFiles.Count) common application files were found. The current user does not have any modification right on them."
-            #         }
-            #     }
-            #     else {
-            #         $Description = "$($Description) This folder does not seem to contain any common application file."
-            #     }
-
-            #     if (($null -ne $RootFolderFiles) -or ($null -ne $RootFolderModifiablePaths)) {
-
-            #         $ModifiableChildPathResult = ($ApplicationFileModifiablePaths | ForEach-Object { Resolve-PathRelativeTo -From $RootFolder.FullName -To $_ } | Select-Object -First $MaxFileCount) -join "; "
-            #         if ($ApplicationFileModifiablePaths.Count -gt $MaxFileCount) { $ModifiableChildPathResult += "; ..." }
-
-            #         $Result = New-Object -TypeName PSObject
-            #         $Result | Add-Member -MemberType "NoteProperty" -Name "Path" -Value $RootFolder.FullName
-            #         $Result | Add-Member -MemberType "NoteProperty" -Name "Modifiable" -Value ($null -ne $RootFolderModifiablePaths)
-            #         $Result | Add-Member -MemberType "NoteProperty" -Name "ModifiablePaths" -Value $ModifiableChildPathResult
-            #         $Result | Add-Member -MemberType "NoteProperty" -Name "Vulnerable" -Value $Vulnerable
-            #         $Result | Add-Member -MemberType "NoteProperty" -Name "Description" -Value $Description
-            #         $AllResults += $Result
-            #     }
-            # }
-
-            $RootFolders | Select-Object -ExpandProperty "Fullname" |
-            Invoke-CommandMultithread -InitialSessionState $(Get-InitialSessionState) -Command "Get-ModifiableRootFolder" -InputParameter "Path" |
-            ForEach-Object { $AllResults += $_ }
+            $Candidates = $RootFolders | Select-Object -ExpandProperty "Fullname"
+            foreach ($Candidate in $Candidates) {
+                Get-ModifiableRootFolder -Path $Candidate | ForEach-Object {
+                    $Results += $_
+                }
+            }
         }
 
-        $Vulnerable = ($AllResults | Where-Object { $_.Vulnerable }).Count -gt 0
+        $Vulnerable = ($Results | Where-Object { $_.Vulnerable }).Count -gt 0
 
         $CheckResult = New-Object -TypeName PSObject
-        $CheckResult | Add-Member -MemberType "NoteProperty" -Name "Result" -Value $AllResults
+        $CheckResult | Add-Member -MemberType "NoteProperty" -Name "Result" -Value $Results
         $CheckResult | Add-Member -MemberType "NoteProperty" -Name "Severity" -Value $(if ($Vulnerable) { $BaseSeverity } else { $script:SeverityLevel::None })
         $CheckResult
     }