+++ src/check/Globals.ps1
@@ -1,6 +1,5 @@
 $script:GlobalVariable = @{
     CheckResultList     = $null
-    InitialSessionState = $null
 }
 
 $script:GlobalCache = @{